Feature/security

+++ src/DataPersister/CustomerDataPersister.php
@@ -17,12 +17,17 @@ final class CustomerDataPersister implements DataPersisterInterface
 
     public function supports($data, array $context = []): bool
     {
+        if ($context['resource_class'] !== Customer::class) {
+            return false;
+        }
         return $data instanceof Customer;
     }
 
     public function persist($data, array $context = [])
     {
-        
+        if ($context['resource_class'] !== Customer::class) {
+            return false;
+        }
         // Set Date 
         $data->setCreatedAt(new \DateTime());
 
@@ -33,6 +38,9 @@ final class CustomerDataPersister implements DataPersisterInterface
 
     public function remove($data, array $context = [])
     {
+        if ($context['resource_class'] !== Customer::class) {
+            return false;
+        }
         // Doctrine Remove Customer
         $this->entityManager->remove($data);
         $this->entityManager->flush();
